requirements.txt
django
gunicorn
django-heroku
#mysqlclient
djangorestframework
markdown
django-filter